requirements/dev.txt
cyclonedx-python-lib==7.5.0
flake8==7.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
jinja2==3.1.4
mypy==1.11.2
pdoc==14.5.0
pre-commit==3.7.1
pyparsing==3.1.2
pytest-cov==5.0.0
pytest==8.2.0
semver==3.0.2
spdx-tools==0.8.3
types-Flask-Migrate
uuid7==0.1.0
sphinx==9.1.0
myst-parser==5.0.0
sphinx-rtd-theme==3.1.0
